+++ commandservice/src/main/java/command/service/repository/CompanyCommandRepository.java
@@ -5,7 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CompanyCommandRepository extends JpaRepository<CompanyCreation, Integer> {
-
+public interface CompanyCommandRepository extends JpaRepository<CompanyCreation, Long> {
 
 }
