+++ commandservice/src/main/java/command/service/repository/StockCommandRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface StockCommandRepository extends JpaRepository<StockCreation, Integer> {
+public interface StockCommandRepository extends JpaRepository<StockCreation, Long> {
 
 }
